# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Buckhill Petshop API",
        "description": "API implementation Buckhill petshop project",
        "contact": {
            "name": "Jugal Kariya",
            "email": "jugalkariya@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/main/blog": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "List all blog posts",
                "operationId": "d46fd327e8065a22b0b7c9fbfaa7256f",
                "responses": {
                    "200": {
                        "description": "A paginated list of blog posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/v1/main/blog?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/v1/main/blog?page=1"
                                                },
                                                "prev": {
                                                    "type": "null",
                                                    "example": null
                                                },
                                                "next": {
                                                    "type": "null",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/v1/main/blog"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/main/blog/{uuid}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a blog post",
                "operationId": "caebfa8b8eec42d2580b41711b6148a8",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the blog post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The blog post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogPost"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Post": {
                "title": "Post API",
                "properties": {
                    "uuid": {
                        "description": "The unique identifier of the blog post",
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "description": "The title of the blog post",
                        "type": "string"
                    },
                    "content": {
                        "description": "The content of the blog post",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time when the blog post was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the blog post was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BlogPost": {
                "properties": {
                    "uuid": {
                        "description": "The unique identifier of the blog post",
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "description": "The title of the blog post",
                        "type": "string"
                    },
                    "content": {
                        "description": "The content of the blog post",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date and time when the blog post was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the blog post was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}